Создал класс ActivityIndicatorView и добавил по объекту этого класса в ContactsViewController и PersonsViewController
